[Tests] Combine K1 and K2 JVM IC tests
There is no test runner implementation for K1 tests, so there
is no need to keep additional abstraction.

#KT-88539
