+++ compiler/incremental-compilation-impl/testFixtures/org/jetbrains/kotlin/incremental/AbstractIncrementalFirICJvmCompilerRunnerTest.kt
@@ -10,14 +10,10 @@ import org.jetbrains.kotlin.config.LanguageVersion
 import org.jetbrains.kotlin.incremental.testingUtils.BuildLogFinder
 import java.io.File
 
-abstract class AbstractIncrementalK2FirICJvmCompilerRunnerTest : AbstractIncrementalJvmCompilerRunnerTest() {
+abstract class AbstractIncrementalFirICJvmCompilerRunnerTest : AbstractIncrementalJvmCompilerRunnerTest() {
     override fun createCompilerArguments(destinationDir: File, testDir: File): K2JVMCompilerArguments =
         super.createCompilerArguments(destinationDir, testDir).apply {
-            languageVersion = LanguageVersion.LATEST_STABLE.versionString
             @Suppress("DEPRECATION")
             useFirIC = true
         }
-
-    override val buildLogFinder: BuildLogFinder
-        get() = BuildLogFinder(isGradleEnabled = true, isFirEnabled = true) // TODO: investigate cases that need isGradleEnabled - the combination looks fragile
 }
