+++ compiler/incremental-compilation-impl/tests-gen/org/jetbrains/kotlin/incremental/IncrementalFirICJvmCompilerRunnerTestGenerated.java
@@ -16,11 +16,11 @@ import java.util.regex.Pattern;
 
 /** This class is generated by {@link org.jetbrains.kotlin.incremental.TestGeneratorForICTestsKt}. DO NOT MODIFY MANUALLY */
 @SuppressWarnings("all")
-public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncrementalK2JvmCompilerRunnerTest {
+public class IncrementalFirICJvmCompilerRunnerTestGenerated extends AbstractIncrementalFirICJvmCompilerRunnerTest {
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/pureKotlin")
   @TestDataPath("$PROJECT_ROOT")
-  public class PureKotlinAbstractIncrementalK2JvmCompilerRunnerTest {
+  public class PureKotlinAbstractIncrementalFirICJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/pureKotlin/" + fileName);
     }
@@ -56,7 +56,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
     }
 
     @Test
-    public void testAllFilesPresentInPureKotlinAbstractIncrementalK2JvmCompilerRunnerTest() {
+    public void testAllFilesPresentInPureKotlinAbstractIncrementalFirICJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/pureKotlin"), Pattern.compile("^([^.]+)$"), Pattern.compile("((^.*Expect.*)|(^removeMemberTypeAlias)|(^addMemberTypeAlias))"), false);
     }
 
@@ -850,13 +850,13 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/classHierarchyAffected")
   @TestDataPath("$PROJECT_ROOT")
-  public class ClassHierarchyAffectedAbstractIncrementalK2JvmCompilerRunnerTest {
+  public class ClassHierarchyAffectedAbstractIncrementalFirICJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/classHierarchyAffected/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInClassHierarchyAffectedAbstractIncrementalK2JvmCompilerRunnerTest() {
+    public void testAllFilesPresentInClassHierarchyAffectedAbstractIncrementalFirICJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/classHierarchyAffected"), Pattern.compile("^([^.]+)$"), null, false);
     }
 
@@ -1116,13 +1116,13 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/inlineFunCallSite")
   @TestDataPath("$PROJECT_ROOT")
-  public class InlineFunCallSiteAbstractIncrementalK2JvmCompilerRunnerTest {
+  public class InlineFunCallSiteAbstractIncrementalFirICJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/inlineFunCallSite/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInInlineFunCallSiteAbstractIncrementalK2JvmCompilerRunnerTest() {
+    public void testAllFilesPresentInInlineFunCallSiteAbstractIncrementalFirICJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/inlineFunCallSite"), Pattern.compile("^([^.]+)$"), null, true);
     }
 
@@ -1430,14 +1430,14 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/withJava")
   @TestDataPath("$PROJECT_ROOT")
-  public class WithJavaAbstractIncrementalK2JvmCompilerRunnerTest {
+  public class WithJavaAbstractIncrementalFirICJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/withJava/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInWithJavaAbstractIncrementalK2JvmCompilerRunnerTest() {
-      KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava"), Pattern.compile("^([^.]+)$"), null, true);
+    public void testAllFilesPresentInWithJavaAbstractIncrementalFirICJvmCompilerRunnerTest() {
+      KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
     }
 
     @Nested
@@ -1450,7 +1450,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
       @Test
       public void testAllFilesPresentInConvertBetweenJavaAndKotlin() {
-        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin"), Pattern.compile("^([^.]+)$"), null, true);
+        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
       }
 
       @Test
@@ -1487,7 +1487,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInJavaToKotlin() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlin"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlin"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1501,7 +1501,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInJavaToKotlinAndBack() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlinAndBack"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlinAndBack"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1515,7 +1515,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInJavaToKotlinAndRemove() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlinAndRemove"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlinAndRemove"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1529,7 +1529,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInKotlinToJava() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/kotlinToJava"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/kotlinToJava"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
     }
@@ -1544,7 +1544,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
       @Test
       public void testAllFilesPresentInJavaUsedInKotlin() {
-        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin"), Pattern.compile("^([^.]+)$"), null, true);
+        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
       }
 
       @Test
@@ -1689,7 +1689,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInChangeFieldType() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeFieldType"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeFieldType"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1703,7 +1703,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInChangeNotUsedSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeNotUsedSignature"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeNotUsedSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1717,7 +1717,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInChangePropertyOverrideType() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changePropertyOverrideType"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changePropertyOverrideType"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1731,7 +1731,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInChangeSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignature"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1745,7 +1745,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInChangeSignaturePackagePrivate() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignaturePackagePrivate"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignaturePackagePrivate"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1759,7 +1759,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInChangeSignaturePackagePrivateNonRoot() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignaturePackagePrivateNonRoot"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignaturePackagePrivateNonRoot"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1773,7 +1773,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInChangeSignatureStatic() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignatureStatic"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignatureStatic"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1787,7 +1787,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInConstantChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantChanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1801,7 +1801,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInConstantPropertyChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantPropertyChanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantPropertyChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1815,7 +1815,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInConstantUnchanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantUnchanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantUnchanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1829,7 +1829,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInEnumEntryAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/enumEntryAdded"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/enumEntryAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1843,7 +1843,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInEnumEntryRemoved() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/enumEntryRemoved"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/enumEntryRemoved"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1857,7 +1857,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInJavaAndKotlinChangedSimultaneously() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaAndKotlinChangedSimultaneously"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaAndKotlinChangedSimultaneously"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1871,7 +1871,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInJavaFieldNullabilityChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaFieldNullabilityChanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaFieldNullabilityChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1885,7 +1885,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInJavaMethodParamNullabilityChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaMethodParamNullabilityChanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaMethodParamNullabilityChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1899,7 +1899,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInJavaMethodReturnTypeNullabilityChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaMethodReturnTypeNullabilityChanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaMethodReturnTypeNullabilityChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1913,7 +1913,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMethodAddedInSuper() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/methodAddedInSuper"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/methodAddedInSuper"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1927,7 +1927,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMethodRenamed() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/methodRenamed"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/methodRenamed"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1941,7 +1941,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMixedInheritance() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/mixedInheritance"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/mixedInheritance"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1955,7 +1955,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInNotChangeSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/notChangeSignature"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/notChangeSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1969,7 +1969,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInRawErrorTypeDuringSerialization() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/rawErrorTypeDuringSerialization"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/rawErrorTypeDuringSerialization"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1983,7 +1983,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInRemoveAnnotation() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/removeAnnotation"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/removeAnnotation"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -1997,7 +1997,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInSamConversions() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
 
         @Test
@@ -2040,7 +2040,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
           @Test
           public void testAllFilesPresentInMethodAddDefault() {
-            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAddDefault"), Pattern.compile("^([^.]+)$"), null, true);
+            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAddDefault"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
           }
         }
 
@@ -2054,7 +2054,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
           @Test
           public void testAllFilesPresentInMethodAdded() {
-            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAdded"), Pattern.compile("^([^.]+)$"), null, true);
+            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
           }
         }
 
@@ -2068,7 +2068,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
           @Test
           public void testAllFilesPresentInMethodAddedSamAdapter() {
-            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAddedSamAdapter"), Pattern.compile("^([^.]+)$"), null, true);
+            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAddedSamAdapter"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
           }
         }
 
@@ -2082,7 +2082,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
           @Test
           public void testAllFilesPresentInMethodSignatureChanged() {
-            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodSignatureChanged"), Pattern.compile("^([^.]+)$"), null, true);
+            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodSignatureChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
           }
         }
 
@@ -2096,7 +2096,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
           @Test
           public void testAllFilesPresentInMethodSignatureChangedSamAdapter() {
-            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodSignatureChangedSamAdapter"), Pattern.compile("^([^.]+)$"), null, true);
+            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodSignatureChangedSamAdapter"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
           }
         }
       }
@@ -2118,7 +2118,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
       @Test
       public void testAllFilesPresentInKotlinUsedInJava() {
-        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava"), Pattern.compile("^([^.]+)$"), null, true);
+        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
       }
 
       @Test
@@ -2215,7 +2215,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInAddOptionalParameter() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/addOptionalParameter"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/addOptionalParameter"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2229,7 +2229,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInChangeNotUsedSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/changeNotUsedSignature"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/changeNotUsedSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2243,7 +2243,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInChangeSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/changeSignature"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/changeSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2257,7 +2257,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInConstantChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/constantChanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/constantChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2271,7 +2271,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInConstantUnchanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/constantUnchanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/constantUnchanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2285,7 +2285,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInFunRenamed() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/funRenamed"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/funRenamed"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2299,7 +2299,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInImportedClassRemoved() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/importedClassRemoved"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/importedClassRemoved"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2313,7 +2313,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInJvmFieldChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/jvmFieldChanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/jvmFieldChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2327,7 +2327,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInJvmFieldUnchanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/jvmFieldUnchanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/jvmFieldUnchanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2341,7 +2341,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMethodAddedInSuper() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/methodAddedInSuper"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/methodAddedInSuper"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2355,7 +2355,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInNotChangeSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/notChangeSignature"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/notChangeSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2369,7 +2369,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInOnlyTopLevelFunctionInFileRemoved() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/onlyTopLevelFunctionInFileRemoved"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/onlyTopLevelFunctionInFileRemoved"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2383,7 +2383,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInPackageFileAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/packageFileAdded"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/packageFileAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2397,7 +2397,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInPrivateChanges() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/privateChanges"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/privateChanges"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2411,7 +2411,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInPropertyRenamed() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/propertyRenamed"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/propertyRenamed"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
     }
@@ -2432,7 +2432,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
       @Test
       public void testAllFilesPresentInOther() {
-        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other"), Pattern.compile("^([^.]+)$"), null, true);
+        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
       }
 
       @Test
@@ -2445,12 +2445,6 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
       @TestMetadata("classRedeclaration")
       public void testClassRedeclaration() {
         runTest("jps/jps-plugin/testData/incremental/withJava/other/classRedeclaration/");
-      }
-
-      @Test
-      @TestMetadata("classToPackageFacade")
-      public void testClassToPackageFacade() {
-        runTest("jps/jps-plugin/testData/incremental/withJava/other/classToPackageFacade/");
       }
 
       @Test
@@ -2613,7 +2607,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInAccessingFunctionsViaRenamedFileClass() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/accessingFunctionsViaRenamedFileClass"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/accessingFunctionsViaRenamedFileClass"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2627,7 +2621,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInAllKotlinFilesRemovedThenNewAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/allKotlinFilesRemovedThenNewAdded"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/allKotlinFilesRemovedThenNewAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2641,7 +2635,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInClassRedeclaration() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/classRedeclaration"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/classRedeclaration"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2655,7 +2649,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInClassToPackageFacade() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/classToPackageFacade"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/classToPackageFacade"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2669,7 +2663,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInDefaultValueInConstructorAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/defaultValueInConstructorAdded"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/defaultValueInConstructorAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2683,7 +2677,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInInlineFunctionWithJvmNameInClass() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineFunctionWithJvmNameInClass"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineFunctionWithJvmNameInClass"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2697,7 +2691,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInInlineTopLevelFunctionWithJvmName() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineTopLevelFunctionWithJvmName"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineTopLevelFunctionWithJvmName"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2711,7 +2705,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInInlineTopLevelValPropertyWithJvmName() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineTopLevelValPropertyWithJvmName"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineTopLevelValPropertyWithJvmName"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2725,7 +2719,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInInnerClassNotGeneratedWhenRebuilding() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/innerClassNotGeneratedWhenRebuilding"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/innerClassNotGeneratedWhenRebuilding"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2739,7 +2733,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInJvmNameChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/jvmNameChanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/jvmNameChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2753,7 +2747,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMainRedeclaration() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/mainRedeclaration"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/mainRedeclaration"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2767,7 +2761,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifileClassAddTopLevelFunWithDefault() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassAddTopLevelFunWithDefault"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassAddTopLevelFunWithDefault"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2781,7 +2775,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifileClassFileAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassFileAdded"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassFileAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2795,7 +2789,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifileClassFileChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassFileChanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassFileChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2809,7 +2803,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifileClassInlineFunction() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassInlineFunction"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassInlineFunction"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2823,7 +2817,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifileClassInlineFunctionAccessingField() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassInlineFunctionAccessingField"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassInlineFunctionAccessingField"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2837,7 +2831,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifileClassRecreated() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRecreated"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRecreated"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2851,7 +2845,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifileClassRecreatedAfterRenaming() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRecreatedAfterRenaming"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRecreatedAfterRenaming"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2865,7 +2859,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifileClassRemoved() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRemoved"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRemoved"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2879,7 +2873,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifileDependantUsage() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileDependantUsage"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileDependantUsage"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2893,7 +2887,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifilePackagePartMethodAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifilePackagePartMethodAdded"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifilePackagePartMethodAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2907,7 +2901,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInMultifilePartsWithProperties() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifilePartsWithProperties"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifilePartsWithProperties"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2921,7 +2915,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInOptionalParameter() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/optionalParameter"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/optionalParameter"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2935,7 +2929,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInPackageFacadeToClass() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageFacadeToClass"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageFacadeToClass"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2949,7 +2943,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInPackageMultifileClassOneFileWithPublicChanges() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageMultifileClassOneFileWithPublicChanges"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageMultifileClassOneFileWithPublicChanges"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2963,7 +2957,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInPackageMultifileClassPrivateOnlyChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageMultifileClassPrivateOnlyChanged"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageMultifileClassPrivateOnlyChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2977,7 +2971,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInPublicPropertyWithPrivateSetterMultiFileFacade() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/publicPropertyWithPrivateSetterMultiFileFacade"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/publicPropertyWithPrivateSetterMultiFileFacade"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -2991,7 +2985,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInTopLevelFunctionWithJvmName() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/topLevelFunctionWithJvmName"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/topLevelFunctionWithJvmName"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
 
@@ -3005,7 +2999,7 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
 
         @Test
         public void testAllFilesPresentInTopLevelPropertyWithJvmName() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/topLevelPropertyWithJvmName"), Pattern.compile("^([^.]+)$"), null, true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/topLevelPropertyWithJvmName"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
         }
       }
     }
@@ -3014,13 +3008,13 @@ public class IncrementalK2JvmCompilerRunnerTestGenerated extends AbstractIncreme
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/incrementalJvmCompilerOnly")
   @TestDataPath("$PROJECT_ROOT")
-  public class IncrementalJvmCompilerOnlyAbstractIncrementalK2JvmCompilerRunnerTest {
+  public class IncrementalJvmCompilerOnlyAbstractIncrementalFirICJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/incrementalJvmCompilerOnly/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInIncrementalJvmCompilerOnlyAbstractIncrementalK2JvmCompilerRunnerTest() {
+    public void testAllFilesPresentInIncrementalJvmCompilerOnlyAbstractIncrementalFirICJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/incrementalJvmCompilerOnly"), Pattern.compile("^([^.]+)$"), null, true);
     }
