+++ compiler/incremental-compilation-impl/tests-gen/org/jetbrains/kotlin/incremental/IncrementalJvmCompilerRunnerTestGenerated.java
@@ -16,11 +16,11 @@ import java.util.regex.Pattern;
 
 /** This class is generated by {@link org.jetbrains.kotlin.incremental.TestGeneratorForICTestsKt}. DO NOT MODIFY MANUALLY */
 @SuppressWarnings("all")
-public class IncrementalK2PsiJvmCompilerRunnerTestGenerated extends AbstractIncrementalK2PsiJvmCompilerRunnerTest {
+public class IncrementalJvmCompilerRunnerTestGenerated extends AbstractIncrementalJvmCompilerRunnerTest {
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/pureKotlin")
   @TestDataPath("$PROJECT_ROOT")
-  public class PureKotlinAbstractIncrementalK2PsiJvmCompilerRunnerTest {
+  public class PureKotlinAbstractIncrementalJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/pureKotlin/" + fileName);
     }
@@ -56,7 +56,7 @@ public class IncrementalK2PsiJvmCompilerRunnerTestGenerated extends AbstractIncr
     }
 
     @Test
-    public void testAllFilesPresentInPureKotlinAbstractIncrementalK2PsiJvmCompilerRunnerTest() {
+    public void testAllFilesPresentInPureKotlinAbstractIncrementalJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/pureKotlin"), Pattern.compile("^([^.]+)$"), Pattern.compile("((^.*Expect.*)|(^removeMemberTypeAlias)|(^addMemberTypeAlias))"), false);
     }
 
@@ -850,13 +850,13 @@ public class IncrementalK2PsiJvmCompilerRunnerTestGenerated extends AbstractIncr
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/classHierarchyAffected")
   @TestDataPath("$PROJECT_ROOT")
-  public class ClassHierarchyAffectedAbstractIncrementalK2PsiJvmCompilerRunnerTest {
+  public class ClassHierarchyAffectedAbstractIncrementalJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/classHierarchyAffected/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInClassHierarchyAffectedAbstractIncrementalK2PsiJvmCompilerRunnerTest() {
+    public void testAllFilesPresentInClassHierarchyAffectedAbstractIncrementalJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/classHierarchyAffected"), Pattern.compile("^([^.]+)$"), null, false);
     }
 
@@ -1116,13 +1116,13 @@ public class IncrementalK2PsiJvmCompilerRunnerTestGenerated extends AbstractIncr
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/inlineFunCallSite")
   @TestDataPath("$PROJECT_ROOT")
-  public class InlineFunCallSiteAbstractIncrementalK2PsiJvmCompilerRunnerTest {
+  public class InlineFunCallSiteAbstractIncrementalJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/inlineFunCallSite/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInInlineFunCallSiteAbstractIncrementalK2PsiJvmCompilerRunnerTest() {
+    public void testAllFilesPresentInInlineFunCallSiteAbstractIncrementalJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/inlineFunCallSite"), Pattern.compile("^([^.]+)$"), null, true);
     }
 
@@ -1430,13 +1430,13 @@ public class IncrementalK2PsiJvmCompilerRunnerTestGenerated extends AbstractIncr
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/withJava")
   @TestDataPath("$PROJECT_ROOT")
-  public class WithJavaAbstractIncrementalK2PsiJvmCompilerRunnerTest {
+  public class WithJavaAbstractIncrementalJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/withJava/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInWithJavaAbstractIncrementalK2PsiJvmCompilerRunnerTest() {
+    public void testAllFilesPresentInWithJavaAbstractIncrementalJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava"), Pattern.compile("^([^.]+)$"), null, true);
     }
 
@@ -3014,13 +3014,13 @@ public class IncrementalK2PsiJvmCompilerRunnerTestGenerated extends AbstractIncr
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/incrementalJvmCompilerOnly")
   @TestDataPath("$PROJECT_ROOT")
-  public class IncrementalJvmCompilerOnlyAbstractIncrementalK2PsiJvmCompilerRunnerTest {
+  public class IncrementalJvmCompilerOnlyAbstractIncrementalJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/incrementalJvmCompilerOnly/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInIncrementalJvmCompilerOnlyAbstractIncrementalK2PsiJvmCompilerRunnerTest() {
+    public void testAllFilesPresentInIncrementalJvmCompilerOnlyAbstractIncrementalJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/incrementalJvmCompilerOnly"), Pattern.compile("^([^.]+)$"), null, true);
     }
