+++ compiler/incremental-compilation-impl/tests-gen/org/jetbrains/kotlin/incremental/IncrementalPsiJvmCompilerRunnerTestGenerated.java
@@ -16,11 +16,11 @@ import java.util.regex.Pattern;
 
 /** This class is generated by {@link org.jetbrains.kotlin.incremental.TestGeneratorForICTestsKt}. DO NOT MODIFY MANUALLY */
 @SuppressWarnings("all")
-public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIncrementalK2FirICJvmCompilerRunnerTest {
+public class IncrementalPsiJvmCompilerRunnerTestGenerated extends AbstractIncrementalPsiJvmCompilerRunnerTest {
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/pureKotlin")
   @TestDataPath("$PROJECT_ROOT")
-  public class PureKotlinAbstractIncrementalK2FirICJvmCompilerRunnerTest {
+  public class PureKotlinAbstractIncrementalPsiJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/pureKotlin/" + fileName);
     }
@@ -56,7 +56,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
     }
 
     @Test
-    public void testAllFilesPresentInPureKotlinAbstractIncrementalK2FirICJvmCompilerRunnerTest() {
+    public void testAllFilesPresentInPureKotlinAbstractIncrementalPsiJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/pureKotlin"), Pattern.compile("^([^.]+)$"), Pattern.compile("((^.*Expect.*)|(^removeMemberTypeAlias)|(^addMemberTypeAlias))"), false);
     }
 
@@ -850,13 +850,13 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/classHierarchyAffected")
   @TestDataPath("$PROJECT_ROOT")
-  public class ClassHierarchyAffectedAbstractIncrementalK2FirICJvmCompilerRunnerTest {
+  public class ClassHierarchyAffectedAbstractIncrementalPsiJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/classHierarchyAffected/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInClassHierarchyAffectedAbstractIncrementalK2FirICJvmCompilerRunnerTest() {
+    public void testAllFilesPresentInClassHierarchyAffectedAbstractIncrementalPsiJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/classHierarchyAffected"), Pattern.compile("^([^.]+)$"), null, false);
     }
 
@@ -1116,13 +1116,13 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/inlineFunCallSite")
   @TestDataPath("$PROJECT_ROOT")
-  public class InlineFunCallSiteAbstractIncrementalK2FirICJvmCompilerRunnerTest {
+  public class InlineFunCallSiteAbstractIncrementalPsiJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/inlineFunCallSite/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInInlineFunCallSiteAbstractIncrementalK2FirICJvmCompilerRunnerTest() {
+    public void testAllFilesPresentInInlineFunCallSiteAbstractIncrementalPsiJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/inlineFunCallSite"), Pattern.compile("^([^.]+)$"), null, true);
     }
 
@@ -1430,14 +1430,14 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/withJava")
   @TestDataPath("$PROJECT_ROOT")
-  public class WithJavaAbstractIncrementalK2FirICJvmCompilerRunnerTest {
+  public class WithJavaAbstractIncrementalPsiJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/withJava/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInWithJavaAbstractIncrementalK2FirICJvmCompilerRunnerTest() {
-      KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+    public void testAllFilesPresentInWithJavaAbstractIncrementalPsiJvmCompilerRunnerTest() {
+      KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava"), Pattern.compile("^([^.]+)$"), null, true);
     }
 
     @Nested
@@ -1450,7 +1450,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
       @Test
       public void testAllFilesPresentInConvertBetweenJavaAndKotlin() {
-        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin"), Pattern.compile("^([^.]+)$"), null, true);
       }
 
       @Test
@@ -1487,7 +1487,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInJavaToKotlin() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlin"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlin"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1501,7 +1501,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInJavaToKotlinAndBack() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlinAndBack"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlinAndBack"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1515,7 +1515,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInJavaToKotlinAndRemove() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlinAndRemove"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/javaToKotlinAndRemove"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1529,7 +1529,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInKotlinToJava() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/kotlinToJava"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/convertBetweenJavaAndKotlin/kotlinToJava"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
     }
@@ -1544,7 +1544,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
       @Test
       public void testAllFilesPresentInJavaUsedInKotlin() {
-        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin"), Pattern.compile("^([^.]+)$"), null, true);
       }
 
       @Test
@@ -1689,7 +1689,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInChangeFieldType() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeFieldType"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeFieldType"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1703,7 +1703,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInChangeNotUsedSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeNotUsedSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeNotUsedSignature"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1717,7 +1717,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInChangePropertyOverrideType() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changePropertyOverrideType"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changePropertyOverrideType"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1731,7 +1731,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInChangeSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignature"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1745,7 +1745,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInChangeSignaturePackagePrivate() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignaturePackagePrivate"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignaturePackagePrivate"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1759,7 +1759,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInChangeSignaturePackagePrivateNonRoot() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignaturePackagePrivateNonRoot"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignaturePackagePrivateNonRoot"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1773,7 +1773,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInChangeSignatureStatic() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignatureStatic"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/changeSignatureStatic"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1787,7 +1787,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInConstantChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantChanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1801,7 +1801,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInConstantPropertyChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantPropertyChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantPropertyChanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1815,7 +1815,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInConstantUnchanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantUnchanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/constantUnchanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1829,7 +1829,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInEnumEntryAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/enumEntryAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/enumEntryAdded"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1843,7 +1843,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInEnumEntryRemoved() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/enumEntryRemoved"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/enumEntryRemoved"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1857,7 +1857,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInJavaAndKotlinChangedSimultaneously() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaAndKotlinChangedSimultaneously"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaAndKotlinChangedSimultaneously"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1871,7 +1871,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInJavaFieldNullabilityChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaFieldNullabilityChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaFieldNullabilityChanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1885,7 +1885,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInJavaMethodParamNullabilityChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaMethodParamNullabilityChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaMethodParamNullabilityChanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1899,7 +1899,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInJavaMethodReturnTypeNullabilityChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaMethodReturnTypeNullabilityChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/javaMethodReturnTypeNullabilityChanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1913,7 +1913,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMethodAddedInSuper() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/methodAddedInSuper"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/methodAddedInSuper"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1927,7 +1927,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMethodRenamed() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/methodRenamed"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/methodRenamed"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1941,7 +1941,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMixedInheritance() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/mixedInheritance"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/mixedInheritance"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1955,7 +1955,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInNotChangeSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/notChangeSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/notChangeSignature"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1969,7 +1969,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInRawErrorTypeDuringSerialization() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/rawErrorTypeDuringSerialization"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/rawErrorTypeDuringSerialization"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1983,7 +1983,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInRemoveAnnotation() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/removeAnnotation"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/removeAnnotation"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -1997,7 +1997,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInSamConversions() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions"), Pattern.compile("^([^.]+)$"), null, true);
         }
 
         @Test
@@ -2040,7 +2040,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
           @Test
           public void testAllFilesPresentInMethodAddDefault() {
-            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAddDefault"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAddDefault"), Pattern.compile("^([^.]+)$"), null, true);
           }
         }
 
@@ -2054,7 +2054,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
           @Test
           public void testAllFilesPresentInMethodAdded() {
-            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAdded"), Pattern.compile("^([^.]+)$"), null, true);
           }
         }
 
@@ -2068,7 +2068,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
           @Test
           public void testAllFilesPresentInMethodAddedSamAdapter() {
-            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAddedSamAdapter"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodAddedSamAdapter"), Pattern.compile("^([^.]+)$"), null, true);
           }
         }
 
@@ -2082,7 +2082,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
           @Test
           public void testAllFilesPresentInMethodSignatureChanged() {
-            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodSignatureChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodSignatureChanged"), Pattern.compile("^([^.]+)$"), null, true);
           }
         }
 
@@ -2096,7 +2096,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
           @Test
           public void testAllFilesPresentInMethodSignatureChangedSamAdapter() {
-            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodSignatureChangedSamAdapter"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+            KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/javaUsedInKotlin/samConversions/methodSignatureChangedSamAdapter"), Pattern.compile("^([^.]+)$"), null, true);
           }
         }
       }
@@ -2118,7 +2118,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
       @Test
       public void testAllFilesPresentInKotlinUsedInJava() {
-        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava"), Pattern.compile("^([^.]+)$"), null, true);
       }
 
       @Test
@@ -2215,7 +2215,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInAddOptionalParameter() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/addOptionalParameter"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/addOptionalParameter"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2229,7 +2229,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInChangeNotUsedSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/changeNotUsedSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/changeNotUsedSignature"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2243,7 +2243,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInChangeSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/changeSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/changeSignature"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2257,7 +2257,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInConstantChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/constantChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/constantChanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2271,7 +2271,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInConstantUnchanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/constantUnchanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/constantUnchanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2285,7 +2285,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInFunRenamed() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/funRenamed"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/funRenamed"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2299,7 +2299,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInImportedClassRemoved() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/importedClassRemoved"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/importedClassRemoved"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2313,7 +2313,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInJvmFieldChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/jvmFieldChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/jvmFieldChanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2327,7 +2327,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInJvmFieldUnchanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/jvmFieldUnchanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/jvmFieldUnchanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2341,7 +2341,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMethodAddedInSuper() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/methodAddedInSuper"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/methodAddedInSuper"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2355,7 +2355,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInNotChangeSignature() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/notChangeSignature"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/notChangeSignature"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2369,7 +2369,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInOnlyTopLevelFunctionInFileRemoved() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/onlyTopLevelFunctionInFileRemoved"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/onlyTopLevelFunctionInFileRemoved"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2383,7 +2383,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInPackageFileAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/packageFileAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/packageFileAdded"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2397,7 +2397,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInPrivateChanges() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/privateChanges"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/privateChanges"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2411,7 +2411,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInPropertyRenamed() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/propertyRenamed"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/kotlinUsedInJava/propertyRenamed"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
     }
@@ -2432,7 +2432,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
       @Test
       public void testAllFilesPresentInOther() {
-        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other"), Pattern.compile("^([^.]+)$"), null, true);
       }
 
       @Test
@@ -2445,6 +2445,12 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
       @TestMetadata("classRedeclaration")
       public void testClassRedeclaration() {
         runTest("jps/jps-plugin/testData/incremental/withJava/other/classRedeclaration/");
+      }
+
+      @Test
+      @TestMetadata("classToPackageFacade")
+      public void testClassToPackageFacade() {
+        runTest("jps/jps-plugin/testData/incremental/withJava/other/classToPackageFacade/");
       }
 
       @Test
@@ -2607,7 +2613,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInAccessingFunctionsViaRenamedFileClass() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/accessingFunctionsViaRenamedFileClass"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/accessingFunctionsViaRenamedFileClass"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2621,7 +2627,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInAllKotlinFilesRemovedThenNewAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/allKotlinFilesRemovedThenNewAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/allKotlinFilesRemovedThenNewAdded"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2635,7 +2641,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInClassRedeclaration() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/classRedeclaration"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/classRedeclaration"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2649,7 +2655,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInClassToPackageFacade() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/classToPackageFacade"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/classToPackageFacade"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2663,7 +2669,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInDefaultValueInConstructorAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/defaultValueInConstructorAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/defaultValueInConstructorAdded"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2677,7 +2683,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInInlineFunctionWithJvmNameInClass() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineFunctionWithJvmNameInClass"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineFunctionWithJvmNameInClass"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2691,7 +2697,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInInlineTopLevelFunctionWithJvmName() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineTopLevelFunctionWithJvmName"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineTopLevelFunctionWithJvmName"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2705,7 +2711,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInInlineTopLevelValPropertyWithJvmName() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineTopLevelValPropertyWithJvmName"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/inlineTopLevelValPropertyWithJvmName"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2719,7 +2725,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInInnerClassNotGeneratedWhenRebuilding() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/innerClassNotGeneratedWhenRebuilding"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/innerClassNotGeneratedWhenRebuilding"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2733,7 +2739,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInJvmNameChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/jvmNameChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/jvmNameChanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2747,7 +2753,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMainRedeclaration() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/mainRedeclaration"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/mainRedeclaration"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2761,7 +2767,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifileClassAddTopLevelFunWithDefault() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassAddTopLevelFunWithDefault"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassAddTopLevelFunWithDefault"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2775,7 +2781,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifileClassFileAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassFileAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassFileAdded"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2789,7 +2795,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifileClassFileChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassFileChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassFileChanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2803,7 +2809,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifileClassInlineFunction() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassInlineFunction"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassInlineFunction"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2817,7 +2823,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifileClassInlineFunctionAccessingField() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassInlineFunctionAccessingField"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassInlineFunctionAccessingField"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2831,7 +2837,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifileClassRecreated() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRecreated"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRecreated"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2845,7 +2851,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifileClassRecreatedAfterRenaming() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRecreatedAfterRenaming"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRecreatedAfterRenaming"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2859,7 +2865,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifileClassRemoved() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRemoved"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileClassRemoved"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2873,7 +2879,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifileDependantUsage() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileDependantUsage"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifileDependantUsage"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2887,7 +2893,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifilePackagePartMethodAdded() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifilePackagePartMethodAdded"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifilePackagePartMethodAdded"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2901,7 +2907,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInMultifilePartsWithProperties() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifilePartsWithProperties"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/multifilePartsWithProperties"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2915,7 +2921,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInOptionalParameter() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/optionalParameter"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/optionalParameter"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2929,7 +2935,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInPackageFacadeToClass() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageFacadeToClass"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageFacadeToClass"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2943,7 +2949,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInPackageMultifileClassOneFileWithPublicChanges() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageMultifileClassOneFileWithPublicChanges"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageMultifileClassOneFileWithPublicChanges"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2957,7 +2963,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInPackageMultifileClassPrivateOnlyChanged() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageMultifileClassPrivateOnlyChanged"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/packageMultifileClassPrivateOnlyChanged"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2971,7 +2977,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInPublicPropertyWithPrivateSetterMultiFileFacade() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/publicPropertyWithPrivateSetterMultiFileFacade"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/publicPropertyWithPrivateSetterMultiFileFacade"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2985,7 +2991,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInTopLevelFunctionWithJvmName() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/topLevelFunctionWithJvmName"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/topLevelFunctionWithJvmName"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
 
@@ -2999,7 +3005,7 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
 
         @Test
         public void testAllFilesPresentInTopLevelPropertyWithJvmName() {
-          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/topLevelPropertyWithJvmName"), Pattern.compile("^([^.]+)$"), Pattern.compile("(^classToPackageFacade)"), true);
+          KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/withJava/other/topLevelPropertyWithJvmName"), Pattern.compile("^([^.]+)$"), null, true);
         }
       }
     }
@@ -3008,13 +3014,13 @@ public class IncrementalK2FirICJvmCompilerRunnerTestGenerated extends AbstractIn
   @Nested
   @TestMetadata("jps/jps-plugin/testData/incremental/incrementalJvmCompilerOnly")
   @TestDataPath("$PROJECT_ROOT")
-  public class IncrementalJvmCompilerOnlyAbstractIncrementalK2FirICJvmCompilerRunnerTest {
+  public class IncrementalJvmCompilerOnlyAbstractIncrementalPsiJvmCompilerRunnerTest {
     private void run(String fileName) {
       runTest("jps/jps-plugin/testData/incremental/incrementalJvmCompilerOnly/" + fileName);
     }
 
     @Test
-    public void testAllFilesPresentInIncrementalJvmCompilerOnlyAbstractIncrementalK2FirICJvmCompilerRunnerTest() {
+    public void testAllFilesPresentInIncrementalJvmCompilerOnlyAbstractIncrementalPsiJvmCompilerRunnerTest() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("jps/jps-plugin/testData/incremental/incrementalJvmCompilerOnly"), Pattern.compile("^([^.]+)$"), null, true);
     }
