+++ plugins/plugin-sandbox/plugin-sandbox-ic-test/tests-gen/org/jetbrains/kotlin/incremental/IncrementalJvmWithPluginCompilerRunnerTestGenerated.java
@@ -17,7 +17,7 @@ import java.util.regex.Pattern;
 @SuppressWarnings("all")
 @TestMetadata("plugins/plugin-sandbox/plugin-sandbox-ic-test/testData/jvmAndKlib/pureKotlin")
 @TestDataPath("$PROJECT_ROOT")
-public class IncrementalK2JvmWithPluginCompilerRunnerTestGenerated extends AbstractIncrementalK2JvmWithPluginCompilerRunnerTest {
+public class IncrementalJvmWithPluginCompilerRunnerTestGenerated extends AbstractIncrementalJvmWithPluginCompilerRunnerTest {
   private void run(String fileName) {
     runTest("plugins/plugin-sandbox/plugin-sandbox-ic-test/testData/jvmAndKlib/pureKotlin/" + fileName);
   }
